+++ .gitignore
@@ -1,0 +1,2 @@
+github_sshkey
+github_sshkey.pub
