+++ custom_driver/006_linux_device_driver_model/README
@@ -1,0 +1,2 @@
+Use method is  matching base on scan id table
+Extend: 006 linux device driver model
