+++ custom_driver/006_linux_device_driver_model/platform.h
@@ -1,0 +1,14 @@
+#ifndef PLATFORM_H
+#define PLATFORM_H
+
+struct private_device_data
+{
+    int size;
+    int perm;
+    const char *serial_number;
+};
+
+#define RDWR 0x11
+#define RDONLY 0x01
+#define WRONLY 0x10
+#endif
